requirements.txt
flask
python-jenkins
requests
humanfriendly
redis
flask-cors
pyyaml
